refactored Participants modal to its own file
