+++ client/src/components/EventCard/ModalMoreDetails/index.jsx
@@ -4,55 +4,10 @@ import React, { useState, useContext, useEffect } from "react";
 import ListGroup from "react-bootstrap/ListGroup";
 import { useNavigate } from "react-router-dom";
 import Button from "react-bootstrap/Button";
-import Accordion from "react-bootstrap/Accordion";
-import AccordionBody from "react-bootstrap/esm/AccordionBody";
 import Modal from "react-bootstrap/Modal";
+import ParticipantModel from "../../ParticipantModal";
 import AuthContext from "../../../services/authContext";
 import { checkIfRequested, getUserByID } from '../../../services'
-
-function ParticipantModel(props) {
-
-  const { attendees } = props;
-  const [show, setShow] = useState(false);
-
-  console.log(attendees)
-  const handleClose = () => setShow(false);
-  const handleShow = () => setShow(true);
-
-  return (
-    <>
-      <Button variant="primary" onClick={handleShow} style={{marginRight: "auto"}}>
-        Participants
-      </Button>
-      <Modal show={show} onHide={handleClose} size="lg">
-        <Modal.Header closeButton>
-          <Modal.Title> Participants </Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          {attendees.length > 0 ?
-          <Accordion>
-          {
-            attendees.map((attendee) => {
-              return (
-                <Accordion.Item eventKey={attendee}>
-                  <Accordion.Header>{attendee.name}</Accordion.Header>
-                  <AccordionBody>
-                    <p>About: {attendee.about}</p>
-                    <p>Email: {attendee.email}</p>
-                    <p>Pronouns: {attendee.pronouns}</p>
-                  </AccordionBody>  
-                </Accordion.Item>
-              )
-            })
-          }
-          </Accordion> :
-          <h3 style={{textAlign: "center"}}>No participants for now</h3>
-          }
-        </Modal.Body>
-      </Modal>
-    </>
-  )
-}
 
 function ModalMoreDetails(props) {
   const { event, requestJoinHandler, modalImageDisplay } = props;
